python-script/requirements.txt
python-hcl2==4.3.4
pyyaml
pytest
